[trace repositories] harmonize configs, little improvements and clean-ups (#16)
We want to have a consistent configuration, for the org's repositories that
we maintain. This means e.g. using the same "branch protection" rules across
them. We've also disabled "merge commits", since we dislike their effect on
the git history, and they were accidently used once in a while, despite us
trying not-to.

Also did some little clean-ups and improvements, e.g. refreshed some of the
repos "topics" and descriptions.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>